New probe that collects cpu and mem usage from docker container (#82)
* Update on the docker probe

Made changes to automate the deployment of a docker probe when it comes to set resourceId and probeId. Added a time period for sending the values being collected by probes. Also added commands to make the probe run in detached mode and to be removed in case it is interrupted. Finally, needed dependices were added to the dockerfile

* Documentation update and descriptionId fix

* Revert "Documentation update and descriptionId fix"

This reverts commit eabc0a2c2f5bcaa6818e792c17a11fd0dfd268f6.

* Documentation update and descriptionId fix

* Add CPU and MEM % docker probe

This probe is a specified version of the docker probe.
Changes to the base version:
-Probe's code (.py file)
-Setup and documentation naming

* Update README.md

* Update README.md

* Fixed Capital letters error

* Update probe-docker-CPU_MEM_usage.py

Added support for collecting the number of containers running an image.
Also, the CPU and MEM % are retrieved with the sum of all containers.

* Update of certificates

* Update of dependicies versions

Updated versions of technologies on Dockerfile, of base image so that monitor component can be deployed without errors

* Update probe-docker-CPU_MEM_usage.py