Update on the docker probe (#80)
* Update on the docker probe

Made changes to automate the deployment of a docker probe when it comes to set resourceId and probeId. Added a time period for sending the values being collected by probes. Also added commands to make the probe run in detached mode and to be removed in case it is interrupted. Finally, needed dependices were added to the dockerfile

* Documentation update and descriptionId fix

* Revert "Documentation update and descriptionId fix"

This reverts commit eabc0a2c2f5bcaa6818e792c17a11fd0dfd268f6.

* Documentation update and descriptionId fix